Asset extracted to separate file, trades index draft
